requirements.txt
# Core
anthropic==0.86.0
litellm==1.82.6
google-genai==1.68.0
python-jobspy==1.1.82
requests==2.32.5
pyyaml==6.0.3

# Database
# (sqlite3 is part of Python standard library — no install needed)

# Web app
flask==3.1.3

# Scheduler
apscheduler==3.11.2

# Utilities
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
waitress==3.0.2

# Profile file format support
pypdf>=4.0          # .pdf profile files
python-docx>=1.1    # .docx profile files

# Testing
pytest>=9.0
